static/tests/samesite/fetch.html

<!DOCTYPE html>
<script src="/tests/resources/testharness.js"></script>
<script src="/tests/resources/testharnessreport.js"></script>
<script src="/tests/resources/cookie-helper.js"></script>
<script>
  var HTTP_COOKIE = "cookie_via_http";

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(document.origin, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Lax")
      .then(_ => {
        return assert_http_cookie(document.origin, HTTP_COOKIE, COOKIE_VALUE, true);
      })
  }, "Same-origin requests are laxly same-site.");

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(document.origin, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Strict")
      .then(_ => {
        return assert_http_cookie(document.origin, HTTP_COOKIE, COOKIE_VALUE, true);
      })
  }, "Same-origin requests are strictly same-site.");

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie("http://127.0.0.1:8080", HTTP_COOKIE, COOKIE_VALUE, "SameSite=Strict")
      .then(_ => {
        return assert_http_cookie("http://127.0.0.1:8080", HTTP_COOKIE, COOKIE_VALUE, false);
      })
  }, "Cross-origin requests are not same-site.");
</script>
